feat: new project layout and new boards support
