+++ .gitignore
@@ -1,0 +1,6 @@
+/.vscode/
+/.west/
+/build/
+/modules/
+/zephyr/
+/zmk/
