+++ CHANGELOG.md
@@ -1,0 +1,23 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Added
+
+- WHKB Pro2 Rev A support.
+
+### Changed
+
+- Upgraded to Zephyr 3.2 and latest ZMK.
+- New project layout based on Zephyr recommendations.
+
+## 1.0.0 - 2022-05-13
+
+Initial release.
+
+[unreleased]: https://github.com/kanru/hhkb-zmk/compare/039f708e5301819adebd99227f8592b7cd955ebe...HEAD
