+++ build.yaml
@@ -13,5 +13,7 @@
 #
 ---
 include:
+  - board: whkb_pro2
+    shield: hhkb_pro2
   - board: nice_nano_v2
-    shield: hhkb
+    shield: hhkb_pro2 custom_pro2
