+++ config/boards/arm/whkb_pro2/Kconfig
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: MIT
+
+config BOARD_ENABLE_DCDC
+    bool "Enable DCDC mode"
+    select SOC_DCDC_NRF52X
+    default y
+    depends on BOARD_WHKB_PRO2
