+++ config/boards/shields/custom_pro2/Kconfig.defconfig
@@ -1,0 +1,6 @@
+if SHIELD_CUSTOM_PRO2
+
+config ZMK_KEYBOARD_NAME
+    default "Custom HHKB Pro2"
+
+endif
