--- config/boards/shields/hhkb/Kconfig.defconfig
@@ -1,6 +1,0 @@
-if SHIELD_HHKB
-
-config ZMK_KEYBOARD_NAME
-    default "HHKB"
-
-endif
