--- config/boards/shields/hhkb/Kconfig.shield
@@ -1,2 +1,0 @@
-config SHIELD_HHKB
-    def_bool $(shields_list_contains,hhkb)
