--- config/boards/shields/hhkb/hhkb.conf
@@ -1,9 +1,0 @@
-CONFIG_BT_MAX_PAIRED=3
-
-CONFIG_ZMK_IDLE_TIMEOUT=5000
-CONFIG_ZMK_SLEEP=y
-CONFIG_ZMK_IDLE_SLEEP_TIMEOUT=900000
-CONFIG_ZMK_BLE=y
-CONFIG_ZMK_EXT_POWER=n
-
-CONFIG_ZMK_BATTERY_REPORT_INTERVAL=3600
