+++ config/boards/shields/hhkb_pro2/hhkb_pro2.keymap
