+++ config/boards/shields/hhkb_pro2/hhkb_pro2.overlay
@@ -1,0 +1,36 @@
+#include <dt-bindings/zmk/matrix_transform.h>
+#include <dt-bindings/zmk/kscan_mock.h>
+
+/ {
+    chosen {
+        zmk,kscan = &kscan0;
+        zmk,matrix_transform = &default_transform;
+    };
+
+    kscan0: kscan0 {
+        compatible = "zmk,kscan-hhkb-pro2";
+        gpios = <&hhkb_pro2_connector 0 GPIO_ACTIVE_HIGH>,               /* 0: POWER */
+                <&hhkb_pro2_connector 1 (GPIO_ACTIVE_LOW|GPIO_PULL_UP)>, /* 1: KEY */
+                <&hhkb_pro2_connector 2 GPIO_ACTIVE_HIGH>,               /* 2: HYS */
+                <&hhkb_pro2_connector 3 GPIO_ACTIVE_HIGH>,               /* 3: BIT0 */
+                <&hhkb_pro2_connector 4 GPIO_ACTIVE_HIGH>,               /* 4: BIT1 */
+                <&hhkb_pro2_connector 5 GPIO_ACTIVE_HIGH>,               /* 5: BIT2 */
+                <&hhkb_pro2_connector 6 GPIO_ACTIVE_HIGH>,               /* 6: BIT3 */
+                <&hhkb_pro2_connector 7 GPIO_ACTIVE_HIGH>,               /* 7: BIT4 */
+                <&hhkb_pro2_connector 8 GPIO_ACTIVE_HIGH>,               /* 8: BIT5 */
+                <&hhkb_pro2_connector 9 GPIO_ACTIVE_HIGH>;               /* 9: STROBE */
+    };
+
+    default_transform: keymap_transform_0 {
+        compatible = "zmk,matrix-transform";
+        columns = <8>;
+        rows = <8>;
+        map = <
+        RC(3,1) RC(3,0) RC(0,0) RC(1,0) RC(1,1) RC(2,0) RC(2,1) RC(4,0) RC(4,1) RC(6,0) RC(6,1) RC(7,0) RC(7,1) RC(5,0) RC(5,1)
+        RC(3,2) RC(0,1) RC(0,2) RC(1,3) RC(1,2) RC(2,3) RC(2,2) RC(4,2) RC(4,3) RC(6,2) RC(6,3) RC(7,3) RC(7,2) RC(5,2)
+        RC(3,3) RC(0,4) RC(0,3) RC(1,4) RC(1,5) RC(2,4) RC(2,5) RC(4,5) RC(4,4) RC(6,5) RC(6,4) RC(7,4) RC(5,3)
+        RC(3,4) RC(0,5) RC(0,6) RC(0,7) RC(1,6) RC(1,7) RC(2,6) RC(4,6) RC(6,6) RC(7,6) RC(7,5) RC(5,5) RC(5,4)
+        RC(3,5) RC(3,6)                 RC(3,7)                 RC(5,7) RC(5,6)
+        >;
+    };
+};
