--- config/hhkb_drivers/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_subdirectory(kscan)
-add_subdirectory(pm)
