--- config/hhkb_drivers/zephyr/module.yml
@@ -1,6 +1,0 @@
-name: hhkb_drivers
-build:
-  depends:
-    - drivers
-  cmake: .
-  kconfig: Kconfig
