+++ docs/images/connector.jpg
